Use patched version of s-bbbosu.adb that prevents race conditions with nested interrupts
This adds a patched version of s-bbbosu.adb that disables interrupts in the
Interrupt_Handler before calling Interrupt_Wrapper. This prevents a race
condition in the thread queue in some rare cases involving nested interrupts.

Note that interrupts are re-enabled later (when safe) in Interrupt_Wrapper
so nested interrupts will still work.